weather-pipeline/requirements-airflow.txt
apache-airflow-providers-google>=8.0.0
apache-airflow-providers-apache-spark>=4.0.0
requests>=2.28.0
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0
